feat(core, fs, cli): complete audit remediations and native Linux interop
- Fix H1: extent bounds checking and hole handling in rimio
- Fix H2-H25: extent uninitialized bit decoding, 64-bit block counts, indirect block mapping, MFT runs, and SFN collisions
- Add strict VBR and superblock validations across FAT, Ext4, ExFAT, and NTFS
- Fix root directory mode preservation on single file injection (prevent S_IFREG overwrite on inode 2)
- Ensure FAT32 FSINFO free cluster and next free cluster updates on flush while preserving RimFAT integrity checksums
- Remove unbounded memory allocations and enforce streaming ceilings across resolvers and injectors
- Resolve all clippy warnings across workspace crates
- Validate native Linux interoperability via WSL (fsck.fat, fsck.ext4, and kernel mounts)
- Update CHANGELOG.md for v0.9.2

+++ rimfs-exfat/src/types/entries.rs
@@ -109,7 +109,7 @@ impl ExFatEntries {
     pub fn file(
         name: &str,
         cluster: u32,
-        size: u32,
+        size: u64,
         attr: &FileAttributes,
         upcase: &UpcaseHandle,
     ) -> FsParsingResult<Self> {
@@ -121,7 +121,7 @@ impl ExFatEntries {
 
         let name_hash = compute_name_hash(name, upcase);
 
-        let mut stream = ExFatStreamEntry::new(cluster, size as u64, name_length, name_hash);
+        let mut stream = ExFatStreamEntry::new(cluster, size, name_length, name_hash);
         if cluster != 0 {
             stream.general_secondary_flags |= 1; // AllocationPossible
         }
@@ -138,7 +138,7 @@ impl ExFatEntries {
     pub fn file_contiguous(
         name: &str,
         cluster: u32,
-        size: u32,
+        size: u64,
         attr: &FileAttributes,
         upcase: &UpcaseHandle,
     ) -> FsParsingResult<Self> {
@@ -150,7 +150,7 @@ impl ExFatEntries {
 
         let name_hash = compute_name_hash(name, upcase);
 
-        let mut stream = ExFatStreamEntry::new(cluster, size as u64, name_length, name_hash);
+        let mut stream = ExFatStreamEntry::new(cluster, size, name_length, name_hash);
         stream.general_secondary_flags |= 0x03; // AllocationPossible | NoFatChain
 
         primary.compute_set_checksum(&stream, &names);
